Enforce 1-script-1-output for changepoints and sensitivity
- Split compute_changepoints.py: remove convergence side-effect.
  New compute_convergence.py reads tab_changepoints.csv, writes
  tab_convergence.csv.
- Replace plot_embedding_sensitivity.py (N figures per invocation)
  with plot_sensitivity.py --projection {pca,jl} (1 figure per call).
- Update divergence.mk: separate targets for each table and figure.
  Every Make invocation now produces exactly one artifact.
